Refactor neovim configuration to be filetype oriented
As a result of having long load times, Neovim's configuration was
refactored so that there is a more sparse default configuration.

In addition to this, I wanted to separate language server mappings so
that later on I can make the descriptions and mappings different on a
per language server level.