fix neovim cursor glitch
There is some odd issue regarding the `cursorcolumn` option in Neovim.
When enabled, sometimes there will be a drawing related glitch that
occurs when typing. Simply put, when you are typing at the top of the
window, there will be something that looks like a cursor that moves at
the bottom of the window. It's distracting, so for now that option
needs to stay disabled.